+++ src/app/db/sql/create-gif-table.sql
