+++ src/app/db/sql/create-gifs-comments-table.sql
