+++ src/app/index.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const bodyParser = require('body-parser');
-const cors = require('../middleware/cors');
+const cors = require('./middleware/cors');
 const dotenv = require('dotenv');
