+++ src/app/providers/databaseprovider.js
@@ -1,4 +1,4 @@
-const DatabaseFactory = require('../../db');
+const DatabaseFactory = require('../db');
 
 module.exports = (c) => {
   c.service('db', (c) => DatabaseFactory.init());
