+++ src/utils/auth/index.js
@@ -4,13 +4,12 @@ const bcrypt = require('bcrypt');
 
 module.exports = () => {
   return {
-    generateAuthToken: async (user) => jwt.sign({ 
-      userId: user.id
-    },
-    process.env.SECRET_KEY || 'secret',
-    { expiresIn: '24h' },
+    generateAuthToken: async (user) => jwt.sign(
+      { userId: user.id },
+      process.env.SECRET_KEY || 'secret',
+      { expiresIn: '24h' },
     ),
-    hashpassword: async (password) => bcrypt.hash(password, 10),
-    authenticate: async (password, hash) => bcrypt.compare(password, hash),
+    hashpassword: (password) => bcrypt.hash(password, 10),
+    authenticate: (password, hash) => bcrypt.compare(password, hash),
   }
 }
